Adding ultrasound test images, handle compressed test images
Adding new images to test the variant dimensions that can be encountered when deidentifying pixel data.  Introduced the ability to handle zipped test files to bypass the github 100Mb file size limit for the RGB cine clip.  Renamed the test file for the pixel dimension tests.
